pyproject.toml
[tool.poetry]
name = "django-rest-api"
version = "0.1.0"
description = ""
authors = ["Lukas <gwon713@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.9"
Django = "3.2.15"
django-mysql = "3.8.1"
djangorestframework = "==3.11.0"
django-redis = "==4.12.1"
mysqlclient = ">=1.3.12"
drf-spectacular = "^0.24.2"
django-extensions = "^3.2.1"
Werkzeug = "^2.2.2"
black = "^22.10.0"
pytest-django = "^4.5.2"
djangorestframework-simplejwt = "^5.2.2"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
